[DX] Extract AbstractStep
