+++ spec/Pim/Bundle/BaseConnectorBundle/Step/ValidatorStepSpec.php
@@ -21,7 +21,7 @@ class ValidatorStepSpec extends ObjectBehavior
     function it_is_a_step()
     {
         $this->shouldHaveType('\Akeneo\Component\Batch\Step\StepInterface');
-        $this->shouldHaveType('\Akeneo\Bundle\BatchBundle\Step\AbstractStep');
+        $this->shouldHaveType('\Akeneo\Component\Batch\Step\AbstractStep');
     }
 
     function it_is_configurable(
