+++ src/Akeneo/Bundle/BatchBundle/Step/ItemStep.php
@@ -9,6 +9,7 @@ use Akeneo\Component\Batch\Item\ItemProcessorInterface;
 use Akeneo\Component\Batch\Item\ItemReaderInterface;
 use Akeneo\Component\Batch\Item\ItemWriterInterface;
 // TODO extract validation to a yml file in the bundle
+use Akeneo\Component\Batch\Step\AbstractStep;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
