+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Step/IncompleteStep.php
@@ -4,13 +4,13 @@ namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Step;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Step\AbstractStep;
+use Akeneo\Component\Batch\Step\AbstractStep;
 
 /**
  * Step used for test and always declared a incomplete execution
  *
  */
-class IncompleteStep extends AbstractStep
+class IncompleteStep extends \Akeneo\Component\Batch\Step\AbstractStep
 {
     /**
      * {@inheritdoc}
