+++ src/Pim/Bundle/EnrichBundle/Connector/Step/MassEditStep.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\EnrichBundle\Connector\Step;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Step\AbstractStep;
+use Akeneo\Component\Batch\Step\AbstractStep;
 use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
 use Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\TemporaryFileCleaner;
 use Pim\Bundle\EnrichBundle\Step\MassEditRemoveTemporaryMediaStep;
@@ -15,7 +15,7 @@ use Pim\Bundle\EnrichBundle\Step\MassEditRemoveTemporaryMediaStep;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MassEditStep extends AbstractStep
+class MassEditStep extends \Akeneo\Component\Batch\Step\AbstractStep
 {
     /** @var array */
     protected $configuration;
