+++ src/Pim/Component/Connector/Step/TaskletStep.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Connector\Step;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Step\AbstractStep;
+use Akeneo\Component\Batch\Step\AbstractStep;
 use Pim\Component\Connector\Step\TaskletInterface;
 
 /**
@@ -11,7 +11,7 @@ use Pim\Component\Connector\Step\TaskletInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TaskletStep extends AbstractStep
+class TaskletStep extends \Akeneo\Component\Batch\Step\AbstractStep
 {
     /** @var array */
     protected $configuration;
