+++ src/Pim/Component/Connector/Step/ValidatorStep.php
@@ -4,7 +4,7 @@ namespace Pim\Component\Connector\Step;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Component\Batch\Item\AbstractConfigurableStepElement;
-use Akeneo\Bundle\BatchBundle\Step\AbstractStep;
+use Akeneo\Component\Batch\Step\AbstractStep;
 use Pim\Component\Connector\Item\CharsetValidator;
 
 /**
